Format: isolate the model-label sanitize boundary
Structure-only commit: format/model_label.rs holds short_model_name
and the sanitize rules that keep personal-environment strings off the
share card — a safety surface (SECURITY.md) that now names itself in
any diff touching it. format/snapshot.rs holds the --snapshot text
rendering; snapshot() tightens to module scope (only snapshot_app was
consumed externally). Number/date formatting stays in format.rs.
